source/requirements.txt
TgCrypto
asyncio
ffmpeg-python
youtube-search-python==1.4.9
youtube-search
httpx==0.13.3
python-dotenv
gitpython
aiofiles
aiohttp
requests
heroku3
psutil
future
wget
speedtest-cli
sqlalchemy
